requirements.txt
alabaster==0.7.11
aniso8601==1.1.0
Babel==2.6.0
brew==0.1.4
certifi==2018.4.16
chardet==3.0.4
click==6.7
docutils==0.14
Flask==0.11.1
Flask-API==1.0
Flask-Cache==0.13.1
Flask-Compress==1.4.0
Flask-JWT==0.2.0
flask-mongoengine==0.7.5
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-PyMongo==0.5.2
Flask-REST==1.1
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-WTF==0.12
idna==2.7
imagesize==1.0.0
itsdangerous==0.24
Jinja2==2.8
kafka==1.3.5
MarkupSafe==0.23
mongoengine==0.10.6
mysqlclient==1.3.12
numpy==1.14.5
packaging==17.1
passlib==1.6.5
pockets==0.6.2
py==1.4.31
Pygments==2.2.0
pymongo==3.3.0
PyMySQL==0.8.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2016.6.1
requests==2.19.1
scikit-learn==0.19.1
scipy==1.1.0
simplejson==3.15.0
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.7.5
sphinxcontrib-napoleon==0.6.1
sphinxcontrib-websupport==1.1.0
urllib3==1.23
uWSGI==2.0.17
Werkzeug==0.11.9
WTForms==2.1
